+++ radical_translations/cms/templatetags/cms_tags.py
@@ -39,7 +39,7 @@ def get_object_crumbs(context) -> list:
         if "agents" in path:
             crumbs[-1]["path"] = f"{crumbs[-1]['path']}{obj.agent_type}s/"
 
-        last_crumb["title"] = obj.title
+        last_crumb["title"] = getattr(obj, "title", str(obj))
     else:
         if "agents" in path:
             crumbs[-1]["path"] = ""