+++ radical_translations/core/documents.py
@@ -33,6 +33,7 @@ copy_to_content = {"copy_to": "content"}
 
 @registry.register_document
 class ResourceDocument(Document):
+    meta = fields.KeywordField()
     content = fields.TextField(attr="title.main_title", store=True)
 
     title = fields.TextField(
@@ -135,6 +136,20 @@ class ResourceDocument(Document):
         ):
             return related_instance.resource
 
+    def prepare_meta(self, instance):
+        if instance.is_original():
+            return "source texts"
+
+        meta = []
+
+        if instance.get_paratext():
+            meta.append("paratexts")
+
+        if instance.is_translation():
+            meta.append("translations")
+
+        return meta
+
     def prepare_title(self, instance):
         titles = [str(instance.title)]
 
@@ -216,6 +231,7 @@ class ResourceDocument(Document):
             for item in instance.classifications.filter(
                 edition__vocabulary__prefix=prefix
             )
+            if item.edition.label.lower() not in ["original", "source-text"]
         ]
 
         for relationship in instance.get_paratext():
@@ -249,7 +265,7 @@ class ResourceDocument(Document):
                 },
                 "roles": [
                     {
-                        "label": f"{role.label} of paratext"
+                        "label": f"{role.label} of translation paratext"
                         if instance.is_paratext()
                         else role.label
                         for role in item.roles.all()