+++ radical_translations/core/vocabularies.py
@@ -158,7 +158,7 @@ class VocabularyTranslationTerms(VocabularyBaseList):
                 "translation of a translation",
             ],
             ["Integral", "Integral"],
-            ["Original", "Original"],
+            ["Source-text", "Source-text"],
             [
                 "Partial",
                 "Partial",
@@ -216,7 +216,7 @@ class VocabularyParatextTerms(VocabularyBaseList):
             ["Epigraph", "Epigraph"],
             ["False imprint date", "False imprint date"],
             ["False imprint", "False imprint"],
-            ["Ficticious imprint", "Ficticious imprint"],
+            ["Ficticious imprint", "Fictitious imprint"],
             [
                 "Hermeneutical",
                 "Hermeneutical",