+++ radical_translations/static/js/search.js
@@ -1,6 +1,9 @@
 const YEAR_MIN = 1516
 const YEAR_MAX = 1820
 
+const META_FACETS = ['meta']
+const RANGE_FACETS = ['year']
+
 Vue.component('bar-chart', {
   extends: VueChartJs.Bar,
   props: ['clickHandler'],
@@ -31,12 +34,18 @@ Vue.component('bar-chart', {
 new Vue({
   el: '#app',
   components: {
+    'l-map': window.Vue2Leaflet.LMap,
+    'l-marker': window.Vue2Leaflet.LMarker,
+    'l-popup': window.Vue2Leaflet.LPopup,
+    'l-tile-layer': window.Vue2Leaflet.LTileLayer,
     'vue-slider': window['vue-slider-component']
   },
   delimiters: ['{[', ']}'],
   data: {
-    url: new URL(`${window.location}api/`),
-    url_suggest: new URL(`${window.location}api/suggest/`),
+    url: new URL(`${window.location.origin}${window.location.pathname}api/`),
+    url_suggest: new URL(
+      `${window.location.origin}${window.location.pathname}api/suggest/`
+    ),
     query: '',
     query_text: '',
     query_dates: [YEAR_MIN, YEAR_MAX],
@@ -53,7 +62,18 @@ new Vue({
     page: 1,
     rangeMarks: (v) => v % 10 === 0,
     data: [],
-    data_suggest: []
+    data_suggest: [],
+    map: {
+      options: {
+        zoomSnap: 0.5
+      },
+      center: window.L.latLng(53.3439, 0),
+      show: false,
+      zoom: 4,
+      url: 'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
+      attribution:
+        '&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
+    }
   },
   watch: {
     query_text: _.debounce(async function () {
@@ -66,14 +86,81 @@ new Vue({
       this.data = await this.search()
     },
     filters: async function (newFilters, oldFilters) {
+      this.page = 1
       this.data = await this.search()
+    },
+    'map.show': function (newShow, oldShow) {
+      if (newShow) {
+        dispatchWindowResizeEvent()
+      }
     }
   },
   created: async function () {
+    this.loadSearchParams()
     this.data = await this.search()
   },
   computed: {
     facets: function () {
+      return this.getFacets().filter(
+        (f) => ![...META_FACETS, ...RANGE_FACETS].includes(f.name)
+      )
+    },
+    metaFacets: function () {
+      const facets = this.getFacets().filter((f) =>
+        META_FACETS.includes(f.name)
+      )
+      if (facets.length == 1) {
+        return facets[0]
+      }
+
+      return facets
+    },
+    rangeFacets: function () {
+      const facets = this.getFacets().filter((f) =>
+        RANGE_FACETS.includes(f.name)
+      )
+
+      return facets
+    },
+    numberOfPages: function () {
+      return Math.ceil(this.data.count / this.data.page_size)
+    },
+    suggestions: function () {
+      let suggestions = []
+
+      if (this.data_suggest !== undefined) {
+        Object.keys(this.data_suggest).forEach((s) => {
+          this.data_suggest[s][0].options.forEach((o) =>
+            suggestions.push(o.text)
+          )
+        })
+      }
+
+      return suggestions
+    }
+  },
+  methods: {
+    getFacetDisplayName: function (name) {
+      return name.replaceAll('_', ' ')
+    },
+    clearFilters: function () {
+      this.filters = []
+      this.page = 1
+      this.query = ''
+      this.query_dates = [YEAR_MIN, YEAR_MAX]
+    },
+    getBucketValue: function (bucket) {
+      return bucket.key_as_string ? bucket.key_as_string : bucket.key
+    },
+    getContributions: function (item) {
+      return item.contributions
+        .filter((c) => c.agent.name !== 'any')
+        .map((c) => ({
+          agent: c.agent,
+          roles: c.roles.filter((r) => r.label !== undefined)
+        }))
+    },
+    getFacets: function () {
       let facets = []
 
       if (this.data.facets !== undefined) {
@@ -83,7 +170,7 @@ new Vue({
           let buckets = this.data.facets[f][name]['buckets']
           let chartData = {
             labels: [],
-            datasets: [{ label: name, backgroundColor: '#9b2923', data: [] }]
+            datasets: [{ label: 'Resources', backgroundColor: '#9b2923', data: [] }]
           }
 
           if (range) {
@@ -124,42 +211,6 @@ new Vue({
 
       return facets
     },
-    numberOfPages: function () {
-      return Math.ceil(this.data.count / this.data.page_size)
-    },
-    suggestions: function () {
-      let suggestions = []
-
-      if (this.data_suggest !== undefined) {
-        Object.keys(this.data_suggest).forEach((s) => {
-          this.data_suggest[s][0].options.forEach((o) =>
-            suggestions.push(o.text)
-          )
-        })
-      }
-
-      return suggestions
-    }
-  },
-  methods: {
-    getFacetDisplayName: function (name) {
-      return name.replaceAll('_', ' ')
-    },
-    clearFilters: function () {
-      this.filters = []
-      this.query = ''
-    },
-    getBucketValue: function (bucket) {
-      return bucket.key_as_string ? bucket.key_as_string : bucket.key
-    },
-    getContributions: function (item) {
-      return item.contributions
-        .filter((c) => c.agent.name !== 'any')
-        .map((c) => ({
-          agent: c.agent,
-          roles: c.roles.filter((r) => r.label !== undefined)
-        }))
-    },
     getFacetCount: function (buckets) {
       return buckets
         .map((el) => el.doc_count)
@@ -195,7 +246,39 @@ new Vue({
         ) !== undefined
       )
     },
+    loadSearchParams: function () {
+      const params = new URLSearchParams(window.location.search)
+
+      let key = 'page'
+      if (params.has(key)) {
+        this.page = params.get(key)
+        params.delete(key)
+      }
+
+      key = 'search'
+      if (params.has(key)) {
+        this.query = params.get(key)
+        params.delete(key)
+      }
+
+      key = 'year__gte'
+      if (params.has(key)) {
+        this.query_dates[0] = params.get(key)
+        params.delete(key)
+      }
+
+      key = 'year__lte'
+      if (params.has(key)) {
+        this.query_dates[1] = params.get(key)
+        params.delete(key)
+      }
+
+      for (const key of params.keys()) {
+        this.updateFilters([key.replace('__term', ''), params.get(key)])
+      }
+    },
     rangeSearch: async function () {
+      this.page = 1
       this.data = await this.search()
     },
     search: async function () {
@@ -227,6 +310,8 @@ new Vue({
       )
 
       this.url.search = params.toString()
+      window.history.pushState({}, '', this.url.search)
+
       const response = await fetch(this.url)
 
       return response.json()