+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Added
 * Data migration to merge duplicate libraries (`Organisation`).
 * Field `radical` to `Agent`.
 * Migration to merge duplicate paratext terms.
+* State management to search.
 
 Changed
 ~~~~~~~