+++ CHANGELOG.rst
@@ -33,6 +33,7 @@ Changed
 * Split filters view into main and secondary.
 * Translation facet label to Translation Terms.
 * Move year facet to main filters section.
+* Order related resources by date, relationship and title.
 
 Fixed
 ~~~~~