+++ CHANGELOG.rst
@@ -23,6 +23,7 @@ Added
 * Help text to both filter types (main and secondary).
 * Help text to remove applied filters.
 * Map view to the search results.
+* Tag to `source text` search results.
 
 Changed
 ~~~~~~~