+++ CHANGELOG.rst
@@ -24,6 +24,7 @@ Added
 * Help text to remove applied filters.
 * Map view to the search results.
 * Tag to `source text` search results.
+* `Paratext` search filter.
 
 Changed
 ~~~~~~~