+++ CHANGELOG.rst
@@ -14,6 +14,10 @@ The format is based on `Keep a Changelog`_, and this project adheres to
 [Unreleased] - yyyy-mm-dd
 --------------------
 
+
+[1.0.0] - 2021-01-27
+--------------------
+
 Added
 ~~~~~
 * Data migration to merge duplicate libraries (`Organisation`).